pytorch2.0_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2022.12.7=pypi_0
charset-normalizer=2.1.1=pypi_0
cmake=3.25.0=pypi_0
cycler=0.11.0=pypi_0
filelock=3.9.0=pypi_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.57.0=pypi_0
idna=3.4=pypi_0
imageio=2.31.1=pypi_0
importlib-metadata=6.8.0=pypi_0
jinja2=3.1.2=pypi_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lit=15.0.7=pypi_0
markdown=3.4.4=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.1.2=pypi_0
mpmath=1.2.1=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.0=pypi_0
numpy=1.24.4=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.1.2.30=pypi_0
openssl=3.0.10=h7f8727e_0
packaging=23.1=pypi_0
pillow=9.4.0=pypi_0
pip=23.2.1=py38h06a4308_0
protobuf=4.24.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyparsing=3.1.1=pypi_0
python=3.8.17=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pywavelets=1.4.1=pypi_0
rawpy=0.18.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.21.0=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.0.0=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tifffile=2023.7.10=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.0.1+cu118=pypi_0
torchaudio=2.0.2+cu118=pypi_0
torchvision=0.15.2+cu118=pypi_0
tqdm=4.66.1=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.4.0=pypi_0
urllib3=1.26.13=pypi_0
werkzeug=2.3.7=pypi_0
wheel=0.38.4=py38h06a4308_0
xz=5.4.2=h5eee18b_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
